split merged views and linkanchors again
