+++ plugins/anchors/linkAnchors.mjs
@@ -26,7 +26,7 @@ function _connectionPerpendicular(view, _magnet, refPoint, opt) {
     var intersections = [];
     if (verticalIntersections) Array.prototype.push.apply(intersections, verticalIntersections);
     if (horizontalIntersections) Array.prototype.push.apply(intersections, horizontalIntersections);
-    if (intersections.length > 0) return refPoint.closestPoint(intersections);
+    if (intersections.length > 0) return refPoint.chooseClosest(intersections);
     if ('fallbackAt' in opt) {
         return getPointAtLink(view, opt.fallbackAt);
     }
@@ -48,7 +48,7 @@ export function resolveRef(fn) {
             if (refView) {
                 if (refView.isNodeConnection(ref)) {
                     var distance = ('fixedAt' in opt) ? opt.fixedAt : '50%';
-                    refPoint = getPointAtLink(view, distance);
+                    refPoint = getPointAtLink(refView, distance);
                 } else {
                     refPoint = refView.getNodeBBox(ref).center();
                 }
