+++ src/cellView.mjs
@@ -66,7 +66,9 @@ export const CellView = View.extend({
         View.call(this, options);
     },
 
-    init: function() {
+    initialize: function() {
+
+        joint.mvc.View.prototype.initialize.apply(this, arguments);
 
         this.cleanNodesCache();
 
@@ -147,9 +149,56 @@ export const CellView = View.extend({
         }
     },
 
-    getBBox: function() {
+    getBBox: function(opt) {
 
-        return this.vel.getBBox({ target: this.paper.svg });
+        var bbox;
+        if (opt && opt.useModelGeometry) {
+            var model = this.model;
+            bbox = model.getBBox().bbox(model.angle());
+        } else {
+            bbox = this.getNodeBBox(this.el);
+        }
+
+        return this.paper.localToPaperRect(bbox);
+    },
+
+    getNodeBBox: function(magnet) {
+
+        var rect = this.getNodeBoundingRect(magnet);
+        var magnetMatrix = this.getNodeMatrix(magnet);
+        var translateMatrix = this.getRootTranslateMatrix();
+        var rotateMatrix = this.getRootRotateMatrix();
+        return V.transformRect(rect, translateMatrix.multiply(rotateMatrix).multiply(magnetMatrix));
+    },
+
+    getNodeUnrotatedBBox: function(magnet) {
+
+        var rect = this.getNodeBoundingRect(magnet);
+        var magnetMatrix = this.getNodeMatrix(magnet);
+        var translateMatrix = this.getRootTranslateMatrix();
+        return V.transformRect(rect, translateMatrix.multiply(magnetMatrix));
+    },
+
+    getRootTranslateMatrix: function() {
+
+        var model = this.model;
+        var position = model.position();
+        var mt = V.createSVGMatrix().translate(position.x, position.y);
+        return mt;
+    },
+
+    getRootRotateMatrix: function() {
+
+        var mr = V.createSVGMatrix();
+        var model = this.model;
+        var angle = model.angle();
+        if (angle) {
+            var bbox = model.getBBox();
+            var cx = bbox.width / 2;
+            var cy = bbox.height / 2;
+            mr = mr.translate(cx, cy).rotate(angle).translate(-cx, -cy);
+        }
+        return mr;
     },
 
     highlight: function(el, opt) {
@@ -266,8 +315,9 @@ export const CellView = View.extend({
         var root = this.el;
         var port = end.port;
         var selector = end.magnet;
+        var model = this.model;
         var magnet;
-        if (port != null && this.model.hasPort(port)) {
+        if (port != null && model.isElement() && model.hasPort(port)) {
             magnet = this.findPortNode(port, selector) || root;
         } else {
             if (!selector) selector = end.selector;
@@ -513,26 +563,6 @@ export const CellView = View.extend({
         var metrics = this.nodeCache(magnet);
         if (metrics.geometryShape === undefined) metrics.geometryShape = V(magnet).toGeometryShape();
         return metrics.geometryShape.clone();
-    },
-
-    getNodeBBox: function(magnet) {
-
-        return V.transformRect(this.getNodeBoundingRect(magnet), this.getNodeMatrix(magnet));
-    },
-
-    getNodeUnrotatedBBox: function(magnet) {
-
-        return this.getNodeBBox(magnet);
-    },
-
-    getRootTranslateMatrix: function() {
-
-        return V.createSVGMatrix();
-    },
-
-    getRootRotateMatrix: function() {
-
-        return V.createSVGMatrix();
     },
 
     isNodeConnection: function(node) {
