+++ src/joint.dia.cell.js
@@ -800,960 +800,3 @@ export const Cell = Backbone.Model.extend({
     }
 });
 
-import { View } from './view.js'
-import {
-    assign,
-    guid,
-    omit,
-    parseDOMJSON,
-    isFunction,
-    isObject,
-    isBoolean,
-    isEmpty,
-    isString,
-    toKebabCase,
-    sortedIndex,
-    merge
-} from './util.js';
-import { Point, Rect } from './geometry.js';
-import V from './vectorizer.js';
-import $ from 'jquery';
-import { ToolsView } from './joint.dia.tools.js'
-
-// CellView base view and controller.
-// --------------------------------------------
-
-// This is the base view and controller for `ElementView` and `LinkView`.
-export const CellView = View.extend({
-
-    tagName: 'g',
-
-    svgElement: true,
-
-    selector: 'root',
-
-    metrics: null,
-
-    className: function() {
-
-        var classNames = ['cell'];
-        var type = this.model.get('type');
-
-        if (type) {
-
-            type.toLowerCase().split('.').forEach(function(value, index, list) {
-                classNames.push('type-' + list.slice(0, index + 1).join('-'));
-            });
-        }
-
-        return classNames.join(' ');
-    },
-
-    attributes: function() {
-        var cell = this.model;
-        return {
-            'model-id': cell.id,
-            'data-type': cell.attributes.type
-        };
-    },
-
-    constructor: function(options) {
-
-        // Make sure a global unique id is assigned to this view. Store this id also to the properties object.
-        // The global unique id makes sure that the same view can be rendered on e.g. different machines and
-        // still be associated to the same object among all those clients. This is necessary for real-time
-        // collaboration mechanism.
-        options.id = options.id || guid(this);
-
-        View.call(this, options);
-    },
-
-    initialize: function() {
-
-        joint.mvc.View.prototype.initialize.apply(this, arguments);
-
-        this.cleanNodesCache();
-
-        // Store reference to this to the <g> DOM element so that the view is accessible through the DOM tree.
-        this.$el.data('view', this);
-
-        this.startListening();
-    },
-
-    startListening: function() {
-        this.listenTo(this.model, 'change', this.onAttributesChange);
-    },
-
-    onAttributesChange: function() {
-        // to be overriden
-    },
-
-    parseDOMJSON: function(markup, root) {
-
-        var doc = parseDOMJSON(markup);
-        var selectors = doc.selectors;
-        var groups = doc.groupSelectors;
-        for (var group in groups) {
-            if (selectors[group]) throw new Error('dia.CellView: ambigious group selector');
-            selectors[group] = groups[group];
-        }
-        if (root) {
-            var rootSelector = this.selector;
-            if (selectors[rootSelector]) throw new Error('dia.CellView: ambiguous root selector.');
-            selectors[rootSelector] = root;
-        }
-        return { fragment: doc.fragment, selectors: selectors };
-    },
-
-    // Return `true` if cell link is allowed to perform a certain UI `feature`.
-    // Example: `can('vertexMove')`, `can('labelMove')`.
-    can: function(feature) {
-
-        var interactive = isFunction(this.options.interactive)
-            ? this.options.interactive(this)
-            : this.options.interactive;
-
-        return (isObject(interactive) && interactive[feature] !== false) ||
-            (isBoolean(interactive) && interactive !== false);
-    },
-
-    findBySelector: function(selector, root, selectors) {
-
-        root || (root = this.el);
-        selectors || (selectors = this.selectors);
-
-        // These are either descendants of `this.$el` of `this.$el` itself.
-        // `.` is a special selector used to select the wrapping `<g>` element.
-        if (!selector || selector === '.') return [root];
-        if (selectors) {
-            var nodes = selectors[selector];
-            if (nodes) {
-                if (Array.isArray(nodes)) return nodes;
-                return [nodes];
-            }
-        }
-        // Maintaining backwards compatibility
-        // e.g. `circle:first` would fail with querySelector() call
-        return $(root).find(selector).toArray();
-    },
-
-    notify: function(eventName) {
-
-        if (this.paper) {
-
-            var args = Array.prototype.slice.call(arguments, 1);
-
-            // Trigger the event on both the element itself and also on the paper.
-            this.trigger.apply(this, [eventName].concat(args));
-
-            // Paper event handlers receive the view object as the first argument.
-            this.paper.trigger.apply(this.paper, [eventName, this].concat(args));
-        }
-    },
-
-    getBBox: function(opt) {
-
-        var bbox;
-        if (opt && opt.useModelGeometry) {
-            var model = this.model;
-            bbox = model.getBBox().bbox(model.angle());
-        } else {
-            bbox = this.getNodeBBox(this.el);
-        }
-
-        return this.paper.localToPaperRect(bbox);
-    },
-
-    getNodeBBox: function(magnet) {
-
-        var rect = this.getNodeBoundingRect(magnet);
-        var magnetMatrix = this.getNodeMatrix(magnet);
-        var translateMatrix = this.getRootTranslateMatrix();
-        var rotateMatrix = this.getRootRotateMatrix();
-        return V.transformRect(rect, translateMatrix.multiply(rotateMatrix).multiply(magnetMatrix));
-    },
-
-    getNodeUnrotatedBBox: function(magnet) {
-
-        var rect = this.getNodeBoundingRect(magnet);
-        var magnetMatrix = this.getNodeMatrix(magnet);
-        var translateMatrix = this.getRootTranslateMatrix();
-        return V.transformRect(rect, translateMatrix.multiply(magnetMatrix));
-    },
-
-    getRootTranslateMatrix: function() {
-
-        var model = this.model;
-        var position = model.position();
-        var mt = V.createSVGMatrix().translate(position.x, position.y);
-        return mt;
-    },
-
-    getRootRotateMatrix: function() {
-
-        var mr = V.createSVGMatrix();
-        var model = this.model;
-        var angle = model.angle();
-        if (angle) {
-            var bbox = model.getBBox();
-            var cx = bbox.width / 2;
-            var cy = bbox.height / 2;
-            mr = mr.translate(cx, cy).rotate(angle).translate(-cx, -cy);
-        }
-        return mr;
-    },
-
-    highlight: function(el, opt) {
-
-        el = !el ? this.el : this.$(el)[0] || this.el;
-
-        // set partial flag if the highlighted element is not the entire view.
-        opt = opt || {};
-        opt.partial = (el !== this.el);
-
-        this.notify('cell:highlight', el, opt);
-        return this;
-    },
-
-    unhighlight: function(el, opt) {
-
-        el = !el ? this.el : this.$(el)[0] || this.el;
-
-        opt = opt || {};
-        opt.partial = el != this.el;
-
-        this.notify('cell:unhighlight', el, opt);
-        return this;
-    },
-
-    // Find the closest element that has the `magnet` attribute set to `true`. If there was not such
-    // an element found, return the root element of the cell view.
-    findMagnet: function(el) {
-
-        var $el = this.$(el);
-        var $rootEl = this.$el;
-
-        if ($el.length === 0) {
-            $el = $rootEl;
-        }
-
-        do {
-
-            var magnet = $el.attr('magnet');
-            if ((magnet || $el.is($rootEl)) && magnet !== 'false') {
-                return $el[0];
-            }
-
-            $el = $el.parent();
-
-        } while ($el.length > 0);
-
-        // If the overall cell has set `magnet === false`, then return `undefined` to
-        // announce there is no magnet found for this cell.
-        // This is especially useful to set on cells that have 'ports'. In this case,
-        // only the ports have set `magnet === true` and the overall element has `magnet === false`.
-        return undefined;
-    },
-
-    // Construct a unique selector for the `el` element within this view.
-    // `prevSelector` is being collected through the recursive call.
-    // No value for `prevSelector` is expected when using this method.
-    getSelector: function(el, prevSelector) {
-
-        var selector;
-
-        if (el === this.el) {
-            if (typeof prevSelector === 'string') selector = '> ' + prevSelector;
-            return selector;
-        }
-
-        if (el) {
-
-            var nthChild = V(el).index() + 1;
-            selector = el.tagName + ':nth-child(' + nthChild + ')';
-
-            if (prevSelector) {
-                selector += ' > ' + prevSelector;
-            }
-
-            selector = this.getSelector(el.parentNode, selector);
-        }
-
-        return selector;
-    },
-
-    getLinkEnd: function(magnet, x, y, link, endType) {
-
-        var model = this.model;
-        var id = model.id;
-        var port = this.findAttribute('port', magnet);
-        // Find a unique `selector` of the element under pointer that is a magnet.
-        var selector = magnet.getAttribute('joint-selector');
-
-        var end = { id: id };
-        if (selector != null) end.magnet = selector;
-        if (port != null) {
-            end.port = port;
-            if (!model.hasPort(port) && !selector) {
-                // port created via the `port` attribute (not API)
-                end.selector = this.getSelector(magnet);
-            }
-        } else if (selector == null && this.el !== magnet) {
-            end.selector = this.getSelector(magnet);
-        }
-
-        var paper = this.paper;
-        var connectionStrategy = paper.options.connectionStrategy;
-        if (typeof connectionStrategy === 'function') {
-            var strategy = connectionStrategy.call(paper, end, this, magnet, new g.Point(x, y), link, endType, paper);
-            if (strategy) end = strategy;
-        }
-
-        return end;
-    },
-
-    getMagnetFromLinkEnd: function(end) {
-
-        var root = this.el;
-        var port = end.port;
-        var selector = end.magnet;
-        var model = this.model;
-        var magnet;
-        if (port != null && model.isElement() && model.hasPort(port)) {
-            magnet = this.findPortNode(port, selector) || root;
-        } else {
-            if (!selector) selector = end.selector;
-            if (!selector && port != null) {
-                // link end has only `id` and `port` property referencing
-                // a port created via the `port` attribute (not API).
-                selector = '[port="' + port + '"]';
-            }
-            magnet = this.findBySelector(selector, root, this.selectors)[0];
-        }
-
-        return magnet;
-    },
-
-    getAttributeDefinition: function(attrName) {
-
-        return this.model.constructor.getAttributeDefinition(attrName);
-    },
-
-    setNodeAttributes: function(node, attrs) {
-
-        if (!isEmpty(attrs)) {
-            if (node instanceof SVGElement) {
-                V(node).attr(attrs);
-            } else {
-                $(node).attr(attrs);
-            }
-        }
-    },
-
-    processNodeAttributes: function(node, attrs) {
-
-        var attrName, attrVal, def, i, n;
-        var normalAttrs, setAttrs, positionAttrs, offsetAttrs;
-        var relatives = [];
-        // divide the attributes between normal and special
-        for (attrName in attrs) {
-            if (!attrs.hasOwnProperty(attrName)) continue;
-            attrVal = attrs[attrName];
-            def = this.getAttributeDefinition(attrName);
-            if (def && (!isFunction(def.qualify) || def.qualify.call(this, attrVal, node, attrs))) {
-                if (isString(def.set)) {
-                    normalAttrs || (normalAttrs = {});
-                    normalAttrs[def.set] = attrVal;
-                }
-                if (attrVal !== null) {
-                    relatives.push(attrName, def);
-                }
-            } else {
-                normalAttrs || (normalAttrs = {});
-                normalAttrs[toKebabCase(attrName)] = attrVal;
-            }
-        }
-
-        // handle the rest of attributes via related method
-        // from the special attributes namespace.
-        for (i = 0, n = relatives.length; i < n; i+=2) {
-            attrName = relatives[i];
-            def = relatives[i+1];
-            attrVal = attrs[attrName];
-            if (isFunction(def.set)) {
-                setAttrs || (setAttrs = {});
-                setAttrs[attrName] = attrVal;
-            }
-            if (isFunction(def.position)) {
-                positionAttrs || (positionAttrs = {});
-                positionAttrs[attrName] = attrVal;
-            }
-            if (isFunction(def.offset)) {
-                offsetAttrs || (offsetAttrs = {});
-                offsetAttrs[attrName] = attrVal;
-            }
-        }
-
-        return {
-            raw: attrs,
-            normal: normalAttrs,
-            set: setAttrs,
-            position: positionAttrs,
-            offset: offsetAttrs
-        };
-    },
-
-    updateRelativeAttributes: function(node, attrs, refBBox, opt) {
-
-        opt || (opt = {});
-
-        var attrName, attrVal, def;
-        var rawAttrs = attrs.raw || {};
-        var nodeAttrs = attrs.normal || {};
-        var setAttrs = attrs.set;
-        var positionAttrs = attrs.position;
-        var offsetAttrs = attrs.offset;
-
-        for (attrName in setAttrs) {
-            attrVal = setAttrs[attrName];
-            def = this.getAttributeDefinition(attrName);
-            // SET - set function should return attributes to be set on the node,
-            // which will affect the node dimensions based on the reference bounding
-            // box. e.g. `width`, `height`, `d`, `rx`, `ry`, `points
-            var setResult = def.set.call(this, attrVal, refBBox.clone(), node, rawAttrs);
-            if (isObject(setResult)) {
-                assign(nodeAttrs, setResult);
-            } else if (setResult !== undefined) {
-                nodeAttrs[attrName] = setResult;
-            }
-        }
-
-        if (node instanceof HTMLElement) {
-            // TODO: setting the `transform` attribute on HTMLElements
-            // via `node.style.transform = 'matrix(...)';` would introduce
-            // a breaking change (e.g. basic.TextBlock).
-            this.setNodeAttributes(node, nodeAttrs);
-            return;
-        }
-
-        // The final translation of the subelement.
-        var nodeTransform = nodeAttrs.transform;
-        var nodeMatrix = V.transformStringToMatrix(nodeTransform);
-        var nodePosition = Point(nodeMatrix.e, nodeMatrix.f);
-        if (nodeTransform) {
-            nodeAttrs = omit(nodeAttrs, 'transform');
-            nodeMatrix.e = nodeMatrix.f = 0;
-        }
-
-        // Calculate node scale determined by the scalable group
-        // only if later needed.
-        var sx, sy, translation;
-        if (positionAttrs || offsetAttrs) {
-            var nodeScale = this.getNodeScale(node, opt.scalableNode);
-            sx = nodeScale.sx;
-            sy = nodeScale.sy;
-        }
-
-        var positioned = false;
-        for (attrName in positionAttrs) {
-            attrVal = positionAttrs[attrName];
-            def = this.getAttributeDefinition(attrName);
-            // POSITION - position function should return a point from the
-            // reference bounding box. The default position of the node is x:0, y:0 of
-            // the reference bounding box or could be further specify by some
-            // SVG attributes e.g. `x`, `y`
-            translation = def.position.call(this, attrVal, refBBox.clone(), node, rawAttrs);
-            if (translation) {
-                nodePosition.offset(Point(translation).scale(sx, sy));
-                positioned || (positioned = true);
-            }
-        }
-
-        // The node bounding box could depend on the `size` set from the previous loop.
-        // Here we know, that all the size attributes have been already set.
-        this.setNodeAttributes(node, nodeAttrs);
-
-        var offseted = false;
-        if (offsetAttrs) {
-            // Check if the node is visible
-            var nodeBoundingRect = this.getNodeBoundingRect(node);
-            if (nodeBoundingRect.width > 0 && nodeBoundingRect.height > 0) {
-                var nodeBBox = V.transformRect(nodeBoundingRect, nodeMatrix).scale(1 / sx, 1 / sy);
-                for (attrName in offsetAttrs) {
-                    attrVal = offsetAttrs[attrName];
-                    def = this.getAttributeDefinition(attrName);
-                    // OFFSET - offset function should return a point from the element
-                    // bounding box. The default offset point is x:0, y:0 (origin) or could be further
-                    // specify with some SVG attributes e.g. `text-anchor`, `cx`, `cy`
-                    translation = def.offset.call(this, attrVal, nodeBBox, node, rawAttrs);
-                    if (translation) {
-                        nodePosition.offset(Point(translation).scale(sx, sy));
-                        offseted || (offseted = true);
-                    }
-                }
-            }
-        }
-
-        // Do not touch node's transform attribute if there is no transformation applied.
-        if (nodeTransform !== undefined || positioned || offseted) {
-            // Round the coordinates to 1 decimal point.
-            nodePosition.round(1);
-            nodeMatrix.e = nodePosition.x;
-            nodeMatrix.f = nodePosition.y;
-            node.setAttribute('transform', V.matrixToTransformString(nodeMatrix));
-            // TODO: store nodeMatrix metrics?
-        }
-    },
-
-    getNodeScale: function(node, scalableNode) {
-
-        // Check if the node is a descendant of the scalable group.
-        var sx, sy;
-        if (scalableNode && scalableNode.contains(node)) {
-            var scale = scalableNode.scale();
-            sx = 1 / scale.sx;
-            sy = 1 / scale.sy;
-        } else {
-            sx = 1;
-            sy = 1;
-        }
-
-        return { sx: sx, sy: sy };
-    },
-
-    cleanNodesCache: function() {
-        this.metrics = {};
-    },
-
-    nodeCache: function(magnet) {
-
-        var metrics = this.metrics;
-        // Don't use cache? It most likely a custom view with overridden update.
-        if (!metrics) return {};
-        var id = V.ensureId(magnet);
-        var value = metrics[id];
-        if (!value) value = metrics[id] = {};
-        return value;
-    },
-
-    getNodeData: function(magnet) {
-
-        var metrics = this.nodeCache(magnet);
-        if (!metrics.data) metrics.data = {};
-        return metrics.data;
-    },
-
-    getNodeBoundingRect: function(magnet) {
-
-        var metrics = this.nodeCache(magnet);
-        if (metrics.boundingRect === undefined) metrics.boundingRect = V(magnet).getBBox();
-        return new g.Rect(metrics.boundingRect);
-    },
-
-    getNodeMatrix: function(magnet) {
-
-        var metrics = this.nodeCache(magnet);
-        if (metrics.magnetMatrix === undefined) {
-            var target = this.rotatableNode || this.el;
-            metrics.magnetMatrix = V(magnet).getTransformToElement(target);
-        }
-        return V.createSVGMatrix(metrics.magnetMatrix);
-    },
-
-    getNodeShape: function(magnet) {
-
-        var metrics = this.nodeCache(magnet);
-        if (metrics.geometryShape === undefined) metrics.geometryShape = V(magnet).toGeometryShape();
-        return metrics.geometryShape.clone();
-    },
-
-    isNodeConnection: function(node) {
-        return this.model.isLink() && (!node || node === this.el);
-    },
-
-    findNodesAttributes: function(attrs, root, selectorCache, selectors) {
-
-        var i, n, nodeAttrs, nodeId;
-        var nodesAttrs = {};
-        var mergeIds = [];
-        for (var selector in attrs) {
-            if (!attrs.hasOwnProperty(selector)) continue;
-            var selected = selectorCache[selector] = this.findBySelector(selector, root, selectors);
-            for (i = 0, n = selected.length; i < n; i++) {
-                var node = selected[i];
-                nodeId = V.ensureId(node);
-                nodeAttrs = attrs[selector];
-                // "unique" selectors are selectors that referencing a single node (defined by `selector`)
-                // groupSelector referencing a single node is not "unique"
-                var unique = (selectors && selectors[selector] === node);
-                var prevNodeAttrs = nodesAttrs[nodeId];
-                if (prevNodeAttrs) {
-                    // Note, that nodes referenced by deprecated `CSS selectors` are not taken into account.
-                    // e.g. css:`.circle` and selector:`circle` can be applied in a random order
-                    if (!prevNodeAttrs.array) {
-                        mergeIds.push(nodeId);
-                        prevNodeAttrs.array = true;
-                        prevNodeAttrs.attributes = [prevNodeAttrs.attributes];
-                        prevNodeAttrs.selectedLength = [prevNodeAttrs.selectedLength];
-                    }
-                    var attributes = prevNodeAttrs.attributes;
-                    var selectedLength = prevNodeAttrs.selectedLength;
-                    if (unique) {
-                        // node referenced by `selector`
-                        attributes.unshift(nodeAttrs);
-                        selectedLength.unshift(-1);
-                    } else {
-                        // node referenced by `groupSelector`
-                        var sortIndex = sortedIndex(selectedLength, n);
-                        attributes.splice(sortIndex, 0, nodeAttrs);
-                        selectedLength.splice(sortIndex, 0, n);
-                    }
-                } else {
-                    nodesAttrs[nodeId] = {
-                        attributes: nodeAttrs,
-                        selectedLength: unique ? -1 : n,
-                        node: node,
-                        array: false
-                    };
-                }
-            }
-        }
-
-        for (i = 0, n = mergeIds.length; i < n; i++) {
-            nodeId = mergeIds[i];
-            nodeAttrs = nodesAttrs[nodeId];
-            nodeAttrs.attributes = merge({}, ...nodeAttrs.attributes.reverse());
-        }
-
-        return nodesAttrs;
-    },
-
-    getEventTarget: function(evt) {
-        // Touchmove/Touchend event's target is not reflecting the element under the coordinates as mousemove does.
-        // It holds the element when a touchstart triggered.
-        var type = evt.type;
-        if (type === 'touchmove' || type === 'touchend') {
-            return document.elementFromPoint(evt.clientX, evt.clientY);
-        }
-
-        return evt.target;
-    },
-
-    // Default is to process the `model.attributes.attrs` object and set attributes on subelements based on the selectors,
-    // unless `attrs` parameter was passed.
-    updateDOMSubtreeAttributes: function(rootNode, attrs, opt) {
-
-        opt || (opt = {});
-        opt.rootBBox || (opt.rootBBox = Rect());
-        opt.selectors || (opt.selectors = this.selectors); // selector collection to use
-
-        // Cache table for query results and bounding box calculation.
-        // Note that `selectorCache` needs to be invalidated for all
-        // `updateAttributes` calls, as the selectors might pointing
-        // to nodes designated by an attribute or elements dynamically
-        // created.
-        var selectorCache = {};
-        var bboxCache = {};
-        var relativeItems = [];
-        var item, node, nodeAttrs, nodeData, processedAttrs;
-
-        var roAttrs = opt.roAttributes;
-        var nodesAttrs = this.findNodesAttributes(roAttrs || attrs, rootNode, selectorCache, opt.selectors);
-        // `nodesAttrs` are different from all attributes, when
-        // rendering only  attributes sent to this method.
-        var nodesAllAttrs = (roAttrs)
-            ? this.findNodesAttributes(attrs, rootNode, selectorCache, opt.selectors)
-            : nodesAttrs;
-
-        for (var nodeId in nodesAttrs) {
-            nodeData = nodesAttrs[nodeId];
-            nodeAttrs = nodeData.attributes;
-            node = nodeData.node;
-            processedAttrs = this.processNodeAttributes(node, nodeAttrs);
-
-            if (!processedAttrs.set && !processedAttrs.position && !processedAttrs.offset) {
-                // Set all the normal attributes right on the SVG/HTML element.
-                this.setNodeAttributes(node, processedAttrs.normal);
-
-            } else {
-
-                var nodeAllAttrs = nodesAllAttrs[nodeId] && nodesAllAttrs[nodeId].attributes;
-                var refSelector = (nodeAllAttrs && (nodeAttrs.ref === undefined))
-                    ? nodeAllAttrs.ref
-                    : nodeAttrs.ref;
-
-                var refNode;
-                if (refSelector) {
-                    refNode = (selectorCache[refSelector] || this.findBySelector(refSelector, rootNode, opt.selectors))[0];
-                    if (!refNode) {
-                        throw new Error('dia.ElementView: "' + refSelector + '" reference does not exist.');
-                    }
-                } else {
-                    refNode = null;
-                }
-
-                item = {
-                    node: node,
-                    refNode: refNode,
-                    processedAttributes: processedAttrs,
-                    allAttributes: nodeAllAttrs
-                };
-
-                // If an element in the list is positioned relative to this one, then
-                // we want to insert this one before it in the list.
-                var itemIndex = relativeItems.findIndex(function(item) {
-                    return item.refNode === node;
-                });
-
-                if (itemIndex > -1) {
-                    relativeItems.splice(itemIndex, 0, item);
-                } else {
-                    relativeItems.push(item);
-                }
-            }
-        }
-
-        var rotatableMatrix;
-        for (var i = 0, n = relativeItems.length; i < n; i++) {
-            item = relativeItems[i];
-            node = item.node;
-            refNode = item.refNode;
-
-            // Find the reference element bounding box. If no reference was provided, we
-            // use the optional bounding box.
-            var vRotatable = V(opt.rotatableNode);
-            var refNodeId = refNode ? V.ensureId(refNode) : '';
-            var isRefNodeRotatable = !!vRotatable && !!refNode && vRotatable.contains(refNode);
-            var unrotatedRefBBox = bboxCache[refNodeId];
-            if (!unrotatedRefBBox) {
-                // Get the bounding box of the reference element relative to the `rotatable` `<g>` (without rotation)
-                // or to the root `<g>` element if no rotatable group present if reference node present.
-                // Uses the bounding box provided.
-                var transformationTarget = (isRefNodeRotatable) ? vRotatable : rootNode;
-                unrotatedRefBBox = bboxCache[refNodeId] = (refNode)
-                    ? V(refNode).getBBox({ target: transformationTarget })
-                    : opt.rootBBox;
-            }
-
-            if (roAttrs) {
-                // if there was a special attribute affecting the position amongst passed-in attributes
-                // we have to merge it with the rest of the element's attributes as they are necessary
-                // to update the position relatively (i.e `ref-x` && 'ref-dx')
-                processedAttrs = this.processNodeAttributes(node, item.allAttributes);
-                this.mergeProcessedAttributes(processedAttrs, item.processedAttributes);
-
-            } else {
-                processedAttrs = item.processedAttributes;
-            }
-
-            var refBBox = unrotatedRefBBox;
-            if (isRefNodeRotatable && !vRotatable.contains(node)) {
-                // if the referenced node is inside the rotatable group while the updated node is outside,
-                // we need to take the rotatable node transformation into account
-                if (!rotatableMatrix) rotatableMatrix = V.transformStringToMatrix(vRotatable.attr('transform'));
-                refBBox = V.transformRect(unrotatedRefBBox, rotatableMatrix);
-            }
-
-            this.updateRelativeAttributes(node, processedAttrs, refBBox, opt);
-        }
-    },
-
-    mergeProcessedAttributes: function(processedAttrs, roProcessedAttrs) {
-
-        processedAttrs.set || (processedAttrs.set = {});
-        processedAttrs.position || (processedAttrs.position = {});
-        processedAttrs.offset || (processedAttrs.offset = {});
-
-        assign(processedAttrs.set, roProcessedAttrs.set);
-        assign(processedAttrs.position, roProcessedAttrs.position);
-        assign(processedAttrs.offset, roProcessedAttrs.offset);
-
-        // Handle also the special transform property.
-        var transform = processedAttrs.normal && processedAttrs.normal.transform;
-        if (transform !== undefined && roProcessedAttrs.normal) {
-            roProcessedAttrs.normal.transform = transform;
-        }
-        processedAttrs.normal = roProcessedAttrs.normal;
-    },
-
-    onRemove: function() {
-        this.removeTools();
-    },
-
-    _toolsView: null,
-
-    hasTools: function(name) {
-        var toolsView = this._toolsView;
-        if (!toolsView) return false;
-        if (!name) return true;
-        return (toolsView.getName() === name);
-    },
-
-    addTools: function(toolsView) {
-
-        this.removeTools();
-
-        if (toolsView instanceof ToolsView) {
-            this._toolsView = toolsView;
-            toolsView.configure({ relatedView: this });
-            toolsView.listenTo(this.paper, 'tools:event', this.onToolEvent.bind(this));
-            toolsView.mount();
-        }
-        return this;
-    },
-
-    updateTools: function(opt) {
-
-        var toolsView = this._toolsView;
-        if (toolsView) toolsView.update(opt);
-        return this;
-    },
-
-    removeTools: function() {
-
-        var toolsView = this._toolsView;
-        if (toolsView) {
-            toolsView.remove();
-            this._toolsView = null;
-        }
-        return this;
-    },
-
-    hideTools: function() {
-
-        var toolsView = this._toolsView;
-        if (toolsView) toolsView.hide();
-        return this;
-    },
-
-    showTools: function() {
-
-        var toolsView = this._toolsView;
-        if (toolsView) toolsView.show();
-        return this;
-    },
-
-    onToolEvent: function(event) {
-        switch (event) {
-            case 'remove':
-                this.removeTools();
-                break;
-            case 'hide':
-                this.hideTools();
-                break;
-            case 'show':
-                this.showTools();
-                break;
-        }
-    },
-
-    // Interaction. The controller part.
-    // ---------------------------------
-
-    // Interaction is handled by the paper and delegated to the view in interest.
-    // `x` & `y` parameters passed to these functions represent the coordinates already snapped to the paper grid.
-    // If necessary, real coordinates can be obtained from the `evt` event object.
-
-    // These functions are supposed to be overriden by the views that inherit from `joint.dia.Cell`,
-    // i.e. `joint.dia.Element` and `joint.dia.Link`.
-
-    pointerdblclick: function(evt, x, y) {
-
-        this.notify('cell:pointerdblclick', evt, x, y);
-    },
-
-    pointerclick: function(evt, x, y) {
-
-        this.notify('cell:pointerclick', evt, x, y);
-    },
-
-    contextmenu: function(evt, x, y) {
-
-        this.notify('cell:contextmenu', evt, x, y);
-    },
-
-    pointerdown: function(evt, x, y) {
-
-        if (this.model.graph) {
-            this.model.startBatch('pointer');
-            this._graph = this.model.graph;
-        }
-
-        this.notify('cell:pointerdown', evt, x, y);
-    },
-
-    pointermove: function(evt, x, y) {
-
-        this.notify('cell:pointermove', evt, x, y);
-    },
-
-    pointerup: function(evt, x, y) {
-
-        this.notify('cell:pointerup', evt, x, y);
-
-        if (this._graph) {
-            // we don't want to trigger event on model as model doesn't
-            // need to be member of collection anymore (remove)
-            this._graph.stopBatch('pointer', { cell: this.model });
-            delete this._graph;
-        }
-    },
-
-    mouseover: function(evt) {
-
-        this.notify('cell:mouseover', evt);
-    },
-
-    mouseout: function(evt) {
-
-        this.notify('cell:mouseout', evt);
-    },
-
-    mouseenter: function(evt) {
-
-        this.notify('cell:mouseenter', evt);
-    },
-
-    mouseleave: function(evt) {
-
-        this.notify('cell:mouseleave', evt);
-    },
-
-    mousewheel: function(evt, x, y, delta) {
-
-        this.notify('cell:mousewheel', evt, x, y, delta);
-    },
-
-    onevent: function(evt, eventName, x, y) {
-
-        this.notify(eventName, evt, x, y);
-    },
-
-    onmagnet: function() {
-
-        // noop
-    },
-
-    magnetpointerdblclick: function() {
-
-        // noop
-    },
-
-    magnetcontextmenu: function() {
-
-        // noop
-    },
-
-    setInteractivity: function(value) {
-
-        this.options.interactive = value;
-    }
-}, {
-
-    dispatchToolsEvent: function(paper, event) {
-        //TODO v.talas es6
-        if ((typeof event === 'string') && (paper instanceof joint.dia.Paper)) {
-            paper.trigger('tools:event', event);
-        }
-    }
-});
