+++ src/linkView.mjs
@@ -66,6 +66,8 @@ export const LinkView = CellView.extend({
         this.metrics = {};
     },
 
+    initFlag: FLAG_RENDER | FLAG_SOURCE | FLAG_TARGET,
+
     presentationAttributes: {
         markup: FLAG_RENDER,
         attrs: FLAG_UPDATE,
@@ -91,7 +93,6 @@ export const LinkView = CellView.extend({
     FLAG_VERTICES: FLAG_VERTICES,
     FLAG_SOURCE: FLAG_SOURCE,
     FLAG_TARGET: FLAG_TARGET,
-    FLAG_INIT: FLAG_RENDER | FLAG_SOURCE | FLAG_TARGET,
 
     onAttributesChange: function(model, opt) {
         var flag = model.getChangeFlag(this.presentationAttributes);
@@ -726,50 +727,59 @@ export const LinkView = CellView.extend({
         };
     },
 
+    findAnchorsOrdered: function(firstEndType, firstRef, secondEndType, secondRef) {
+
+        var firstAnchor, secondAnchor;
+        var firstAnchorRef, secondAnchorRef;
+        var model = this.model;
+        var firstDef = model.get(firstEndType);
+        var secondDef = model.get(secondEndType);
+        var firstView = this.getEndView(firstEndType);
+        var secondView = this.getEndView(secondEndType);
+        var firstMagnet = this.getEndMagnet(firstEndType);
+        var secondMagnet = this.getEndMagnet(secondEndType);
+
+        // Anchor first
+        if (firstView) {
+            if (firstRef) {
+                firstAnchorRef = new g.Point(firstRef);
+            } else if (secondView) {
+                firstAnchorRef = secondMagnet;
+            } else {
+                firstAnchorRef = new g.Point(secondDef);
+            }
+            firstAnchor = this.getAnchor(firstDef.anchor, firstView, firstMagnet, firstAnchorRef, firstEndType);
+        } else {
+            firstAnchor = new g.Point(firstDef);
+        }
+
+        // Anchor second
+        if (secondView) {
+            secondAnchorRef = new g.Point(secondRef || firstAnchor);
+            secondAnchor = this.getAnchor(secondDef.anchor, secondView, secondMagnet, secondAnchorRef, secondEndType);
+        } else {
+            secondAnchor = new g.Point(secondDef);
+        }
+
+        var res = {};
+        res[firstEndType] = firstAnchor;
+        res[secondEndType] = secondAnchor;
+        return res;
+    },
+
     findAnchors: function(vertices) {
 
         var model = this.model;
         var firstVertex = vertices[0];
         var lastVertex = vertices[vertices.length - 1];
-        var sourceDef = model.get('source');
-        var targetDef = model.get('target');
-        var sourceView = this.sourceView;
-        var targetView = this.targetView;
-        var sourceMagnet, targetMagnet;
 
-        // Anchor Source
-        var sourceAnchor;
-        if (sourceView) {
-            sourceMagnet = (this.sourceMagnet || sourceView.el);
-            var sourceAnchorRef;
-            if (firstVertex) {
-                sourceAnchorRef = new Point(firstVertex);
-            } else if (targetView) {
-                // TODO: the source anchor reference is not a point, how to deal with this?
-                sourceAnchorRef = this.targetMagnet || targetView.el;
-            } else {
-                sourceAnchorRef = new Point(targetDef);
-            }
-            sourceAnchor = this.getAnchor(sourceDef.anchor, sourceView, sourceMagnet, sourceAnchorRef, 'source');
-        } else {
-            sourceAnchor = new Point(sourceDef);
+        if (model.target().priority && !model.source().priority) {
+            // Reversed order
+            return this.findAnchorsOrdered('target', lastVertex, 'source', firstVertex);
         }
 
-        // Anchor Target
-        var targetAnchor;
-        if (targetView) {
-            targetMagnet = (this.targetMagnet || targetView.el);
-            var targetAnchorRef = new Point(lastVertex || sourceAnchor);
-            targetAnchor = this.getAnchor(targetDef.anchor, targetView, targetMagnet, targetAnchorRef, 'target');
-        } else {
-            targetAnchor = new Point(targetDef);
-        }
-
-        // Con
-        return {
-            source: sourceAnchor,
-            target: targetAnchor
-        };
+        // Usual order
+        return this.findAnchorsOrdered('source', firstVertex, 'target', lastVertex);
     },
 
     findConnectionPoints: function(route, sourceAnchor, targetAnchor) {
@@ -872,7 +882,7 @@ export const LinkView = CellView.extend({
             if (typeof connectionPointFn !== 'function') throw new Error('Unknown connection point: ' + connectionPointName);
         }
         connectionPoint = connectionPointFn.call(this, line, view, magnet, connectionPointDef.args || {}, endType, this);
-        if (!connectionPoint) anchor;
+        if (!connectionPoint) return anchor;
         return connectionPoint.round(this.decimalsRounding);
     },
 
@@ -1342,7 +1352,7 @@ export const LinkView = CellView.extend({
 
         var path = this.path;
         if (!path) return null;
-
+        if (joint.util.isPercentage(ratio)) ratio = parseFloat(ratio) / 100;
         return path.pointAt(ratio, { segmentSubdivisions: this.getConnectionSubdivisions() });
     },
 
@@ -2271,7 +2281,7 @@ Object.defineProperty(LinkView.prototype, 'sourceBBox', {
         }
         var sourceMagnet = this.sourceMagnet;
         if (sourceView.isNodeConnection(sourceMagnet)) {
-            return new Rect(this.sourcePoint);
+            return new Rect(this.sourceAnchor);
         }
         return sourceView.getNodeBBox(sourceMagnet || sourceView.el);
     }
@@ -2290,7 +2300,7 @@ Object.defineProperty(LinkView.prototype, 'targetBBox', {
         }
         var targetMagnet = this.targetMagnet;
         if (targetView.isNodeConnection(targetMagnet)) {
-            return new Rect(this.targetPoint);
+            return new Rect(this.targetAnchor);
         }
         return targetView.getNodeBBox(targetMagnet || targetView.el);
     }
